+++ deno.jsonc
@@ -25,7 +25,7 @@
   "include": ["."],
   "tasks": {
     "run": "deno run -A --no-check=remote",
-    "run:browser": "deno task run test_ctx.ts deno task run _tasks/run_browser.ts",
+    "run:browser": "deno task run test_util/ctx.ts deno task run _tasks/run_browser.ts",
     "debug": "deno task run --inspect-brk",
     "download:frame_metadata": "deno task run _tasks/download_frame_metadata.ts",
     "build:wasm": "deno task run _tasks/build_wasm.ts",
@@ -34,7 +34,7 @@
     "star": "deno task run _tasks/star.ts && deno cache --no-check=remote _star.ts",
     "lock": "deno task star --lock=lock.json --lock-write",
     "lint": "deno lint",
-    "test": "deno task run test_ctx.ts deno test -A --no-check=remote -L=info --ignore=target --parallel",
+    "test": "deno task run test_util/ctx.ts deno test -A -L=info --ignore=target --parallel",
     "test:update": "deno task test -- -- --update",
     "mdbook:watch": "mdbook watch -o",
     "bench": "deno bench -A --no-check=remote --unstable"
